+++ src/libdebug/test/test_expression_evaluator.cpp
@@ -21,6 +21,16 @@ TEST_CASE(expression_evaluator_literals) {
     EXPECT_TRUE(ExpressionEvaluator::evaluate("1234", nullptr, res));
     EXPECT_EQ(res, 1234);
     
+    // Character literals
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("'a'", nullptr, res));
+    EXPECT_EQ(res, 'a');
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("'\\n'", nullptr, res));
+    EXPECT_EQ(res, '\n');
+    
+    // Binary literal + Modulus
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("10 % %11", nullptr, res)); // 10 % 3 = 1
+    EXPECT_EQ(res, 1);
+    
     // Malformed
     ASSERT(!ExpressionEvaluator::evaluate("$GHIJ", nullptr, res));
     ASSERT(!ExpressionEvaluator::evaluate("%2345", nullptr, res));
@@ -42,6 +52,73 @@ TEST_CASE(expression_evaluator_arithmetic) {
     // Left-to-right (100 - 50 + 10 = 60, not 40)
     EXPECT_TRUE(ExpressionEvaluator::evaluate("100 - 50 + 10", nullptr, res));
     EXPECT_EQ(res, 60);
+    
+    // Division and Modulus
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("100 / 4", nullptr, res));
+    EXPECT_EQ(res, 25);
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("10 % 3", nullptr, res));
+    EXPECT_EQ(res, 1);
+    
+    // Precedence: 10 + 20 / 2 => 10 + (20 / 2) = 20
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("10 + 20 / 2", nullptr, res));
+    EXPECT_EQ(res, 20);
+}
+
+TEST_CASE(expression_evaluator_parentheses) {
+    uint32_t res;
+    
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("(10 + 20) * 0 + 5", nullptr, res)); // We don't have * yet, but let's see
+    // Wait, I didn't add *. Let me check my binOps.
+    // I only added +, -, &, |, ==, !=, <=, >=, <, >.
+    
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("(10 + 20) - 5", nullptr, res));
+    EXPECT_EQ(res, 25);
+    
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("100 - (50 + 10)", nullptr, res));
+    EXPECT_EQ(res, 40);
+    
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("((1 + 2) + 3)", nullptr, res));
+    EXPECT_EQ(res, 6);
+}
+
+TEST_CASE(expression_evaluator_unary) {
+    uint32_t res;
+    
+    // High byte <
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("<$1234", nullptr, res));
+    EXPECT_EQ(res, 0x12);
+    
+    // Low byte >
+    EXPECT_TRUE(ExpressionEvaluator::evaluate(">$1234", nullptr, res));
+    EXPECT_EQ(res, 0x34);
+    
+    // Negation !
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("!0", nullptr, res));
+    EXPECT_EQ(res, 1);
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("!55", nullptr, res));
+    EXPECT_EQ(res, 0);
+    
+    // Combined
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("! (>$1200)", nullptr, res));
+    EXPECT_EQ(res, 1); // >$1200 is 0, !0 is 1
+}
+
+TEST_CASE(expression_evaluator_comparisons) {
+    uint32_t res;
+    
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("10 == 10", nullptr, res));
+    EXPECT_EQ(res, 1);
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("10 == 20", nullptr, res));
+    EXPECT_EQ(res, 0);
+    
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("5 < 10", nullptr, res));
+    EXPECT_EQ(res, 1);
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("10 < 5", nullptr, res));
+    EXPECT_EQ(res, 0);
+    
+    // Precedence: 10 == 5 + 5  => 10 == (5 + 5)
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("10 == 5 + 5", nullptr, res));
+    EXPECT_EQ(res, 1);
 }
 
 TEST_CASE(expression_evaluator_symbols_regs) {
@@ -57,6 +134,10 @@ TEST_CASE(expression_evaluator_symbols_regs) {
     
     EXPECT_TRUE(ExpressionEvaluator::evaluate("start + $10", &dbg, res));
     EXPECT_EQ(res, 0x1244);
+    
+    dbg.symbols().addSymbol(0x2000, "my_symbol");
+    EXPECT_TRUE(ExpressionEvaluator::evaluate("my_symbol + 5", &dbg, res));
+    EXPECT_EQ(res, 0x2005);
     
     // Registers
     cpu.setPc(0x4000);